+++ index.html
@@ -5,7 +5,7 @@
   <meta charset="UTF-8" />
   <link rel="icon" type="image/svg+xml" href="/src/static/favicon.svg" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <meta name="description" content="A tool to help you narrow down possible Ghosts on your ghost hunt. Filter by confirmed and excluded evidence." />
+  <meta name="description" content="A tool to help you narrow down possible Ghosts on your ghost hunt. Filter by confirmed and excluded evidence. Better than the other Ghost Filter" />
   <meta name="keywords" content="Phasmophobia Ghost Filter evidence include exclude narrow updated emt spirit box fingerprints ghost orbs writing freezing temperatures temp DOTS D.O.T.S.">
   <title>Phasmophobia Ghost Filter</title>
   <script async src="https://www.googletagmanager.com/gtag/js?id=G-JXZ4TCSBXG"></script>
