+++ src/components/App.module.scss
@@ -1,4 +1,4 @@
-@use "variables";
+@use "../styles/variables";
 
 .App {
   min-height: 100vh;
