+++ src/components/App.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { useRecoilValue } from 'recoil'
 import '../styles/global.scss'
 import { darkModeState } from '../utils/state'
-import styles from '../styles/App.module.scss'
+import styles from './App.module.scss'
 import GhostFilter from './GhostFilter'
 import GhostList from './GhostList'
 import GithubMark from './GithubMark'
