+++ src/components/DarkModeToggle.module.scss
@@ -1,4 +1,4 @@
-@use 'variables';
+@use "../styles/variables";
 
 .darkModeToggle {
   position: relative;
