+++ src/components/DarkModeToggle.tsx
@@ -2,9 +2,9 @@ import React from 'react'
 import { useRecoilState } from 'recoil'
 import clsx from 'clsx'
 import { darkModeState } from '../utils/state'
-import styles from '../styles/DarkModeToggle.module.scss'
 import SunIcon from '../static/sun.svg?component'
 import MoonIcon from '../static/moon.svg?component'
+import styles from './DarkModeToggle.module.scss'
 
 const DarkModeToggle = () => {
   const [darkMode, setDarkMode] = useRecoilState(darkModeState)
