+++ src/components/Ghost.module.scss
@@ -1,5 +1,5 @@
-@use 'sass:math';
-@use 'variables';
+@use "sass:math";
+@use "../styles/variables";
 
 .ghost {
   flex: 1;
