+++ src/components/Ghost.tsx
@@ -4,7 +4,7 @@ import { useRecoilValue } from 'recoil'
 import { Evidence, GhostData } from '../data/ghostData'
 import { isAnyEvidenceSelectedState, possibleGhostsState, possibleRemainingEvidenceState } from '../utils/state'
 import { evidencePrettyName } from '../utils/utils'
-import styles from '../styles/Ghost.module.scss'
+import styles from './Ghost.module.scss'
 import LabelWithCB from './LabelWithCB'
 
 type GhostProps = {
