+++ src/components/GhostList.module.scss
