+++ src/components/LabelWithCB.module.scss
@@ -1,4 +1,4 @@
-@use "mixins";
+@use "../styles/mixins";
 
 .labelWithCB {
   label {
