+++ src/styles/_variables.scss
@@ -1,4 +1,5 @@
 // App width
+// TODO: responsiveness
 $appWidth: 1400px;
 // Transition speed for animations (dark mode switch etc)
 $transitionSpeed: 0.2s;
